+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 **Added**
 
+- #555 Don't allow the deactivation Analysis Services with active dependencies
+- #555 Don't allow the activation of Analysis Services with inactive dependents
 
 **Removed**